20: Moved MobileMenu to Page and other styling improvements
