+++ src/components/WalletInfo/WalletInfo.scss
@@ -1,16 +1,18 @@
+@import "src/styles";
+
 .wallet-info {
   display: flex;
   align-items: center;
   padding: 1rem;
 
   &--is-menu {
-    background-color: var(--c-grey);
+    background-color: var(--c-dark);
     border: 1px solid var(--c-dark-border);
     border-radius: .3125rem;
   }
 
   &--is-banner {
-    background-color: transparent;
+    background-color: var(--c-grey);
   }
    
   &__img { 
@@ -20,6 +22,8 @@
   }
 
   &__address {
+    @extend %text-ellipsis;
+
     margin-left: 0.875rem;
     flex-grow: 1;
     font-weight: var(--fw-700);
@@ -27,6 +31,7 @@
   
   &__icon {
     color: var(--c-white);
-    height: 1.125rem;
+    width: 1.25rem;
+    height: 1.25rem;
   }
 }
