+++ src/compositions/Page/Page.scss
@@ -5,6 +5,23 @@
     margin: 0 auto;
     max-width: var(--page-max-width);
 
+    &__top-bar {
+        z-index: 3;
+    }
+
+    &__mobile-menu {
+        position: fixed;
+        left: 0;
+        top: var(--top-bar-height);
+        width: 100vw;
+        height: calc(100svh - var(--top-bar-height));
+        z-index: 2;
+
+        @include for-size(tablet-portrait-up) {
+            display: none;
+        }
+    }
+
     &__content {
         position: relative;
         padding: var(--page-padding);
