--- src/compositions/TopBar/MobileMenu/MobileMenu.scss
@@ -1,21 +1,0 @@
-.mobile-menu {
-    display: flex;
-    flex-direction: column;
-    padding: 1.5rem;
-    font-weight: var(--fw-700);
-    background-color: var(--c-black);
-    transform: translateY(0vh);
-    transition: transform 0.5s ease-out;
-    overflow-y: auto;
-    
-    &--is-hidden {
-        transform: translateY(-100vh);
-    }
-
-    &__nav-link {
-        display: flex;
-        align-items: center;
-        margin-top: 2.5rem;
-        padding-left: 1rem;
-    }
-}
