+++ src/compositions/TopBar/TopBar.tsx
@@ -1,28 +1,23 @@
-import React, { FC, useState } from 'react';
+import React, { FC } from 'react';
 
 import classNames from 'classnames';
 
 import { AppRoutes } from '../../routes';
 import IconButton from '../IconButton/IconButton';
 import IconNavLink from '../IconNavLink/IconNavLink';
-import MobileMenu from './MobileMenu/MobileMenu';
 
 import './TopBar.scss';
 
 interface TopBarProps {
+  mobileMenuIsVisible: boolean;
+  onMobileMenuButtonClick: () => void;
   className?: string;
 }
 
-const TopBar: FC<TopBarProps> = ({ className = '' }) => {
-  const [mobileMenuIsVisible, setMobileMenuIsVisible] = useState<boolean>(false);
-
+const TopBar: FC<TopBarProps> = ({ mobileMenuIsVisible, onMobileMenuButtonClick, className = '' }) => {
   const containerClassName = classNames('top-bar', {
     'top-bar--mobile-menu-is-visible': mobileMenuIsVisible,
   }, className);
-
-  const handleIconButtonClick = (): void => {
-    setMobileMenuIsVisible(!mobileMenuIsVisible);
-  };
 
   return (
     <div className={containerClassName}>
@@ -36,7 +31,7 @@ const TopBar: FC<TopBarProps> = ({ className = '' }) => {
         hideLabel
         icon={!mobileMenuIsVisible ? 'menu' : 'close'}
         text="Menu button"
-        onClick={handleIconButtonClick}
+        onClick={onMobileMenuButtonClick}
         className="top-bar__menu-button"
         iconClassName="top-bar__menu-button-icon"
       />
@@ -47,7 +42,6 @@ const TopBar: FC<TopBarProps> = ({ className = '' }) => {
         className="top-bar__list-button"
         iconClassName="top-bar__list-button-icon"
       />
-      <MobileMenu isHidden={!mobileMenuIsVisible} className="top-bar__mobile-menu" />
     </div>
   );
 };
