+++ src/styles/extends/button.scss
@@ -4,23 +4,18 @@
     @extend %flex-align-center;
 
     margin: 0;
-    padding: .25rem;
+    padding: 1rem;
     font: inherit;
     font-size: 1rem;
     font-weight: 500;
     line-height: 1;
     color: var(--c-primary);
     cursor: pointer;
-    border: none;
     background: none;
 
     &:hover,
     &:focus {
         outline: none;
-    }
-
-    &:focus {
-        border: 1px solid var(--c-primary);
     }
 
     &:disabled {
